bot/requirements.txt
httpx
python-dotenv
twitchio